server/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
decorator==4.3.0
Flask==1.0.2
future==0.17.1
geocoder==1.38.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
ratelim==0.1.6
requests==2.20.1
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
gunicorn==19.9.0
fenixedu==1.0.0
google-cloud-datastore==1.7.3
python-binary-memcached==0.28.0
numpy==1.15.4
pika==0.12.0
google-cloud-pubsub==0.39.1
grpc-google-iam-v1==0.11.4
firebase-admin==2.14.0
geopy==1.18.1
